requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.0
selenium>=4.11.0
webdriver-manager>=4.0.0
python-multipart>=0.0.6
pillow>=9.0.0 